Remove references to the old "Crunchtime" version of the app (#10)
When we were trying to white-label this as "CrunchTime", we had a setup where the backend would return some configuration values to the frontend like the color scheme and university name. That isn't the plan any more, so removing it.

This also removes references to an API proxy in the application.properties file that is no longer relevant.