Fixed bug
Now, the min and max are connected to their own "check" variable, which is not directly connected to the api call. So when the min and max are changed, on event blur or pressing enter, it will run the check function for the following criteria: if min < max, if min and max are integers, if min and max are numbers at all, and if min < 0.